% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{inference}
\alias{inference}
\title{Given a test market, analyze the impact of an intervention
\code{inference} Analyzes the causal impact of an intervention using the CausalImpact package, given a test market and a matched_market object from the best_matches function.
The function returns an object of type "market_inference" which contains the estimated impact of the intervention (absolute and relative).}
\usage{
inference(matched_markets = NULL, test_market = NULL,
  end_post_period = NULL, alpha = 0.05, prior_level_sd = 0.01,
  control_matches = 5)
}
\arguments{
\item{matched_markets}{A matched_market object created by the market_matching function}

\item{test_market}{The name of the test market (character)}

\item{end_post_period}{The end date of the post period. Must be a character of format "YYYY-MM-DD" -- e.g., "2015-11-01"}

\item{alpha}{Desired tail-area probability for posterior intervals. For example, 0.05 yields 0.95 intervals}

\item{prior_level_sd}{Prior SD for the local level term. Default is 0.01. The biggesr this number is, the more wiggliness is allowed for the local level term.
Note that more wiggly local level terms also translate into larger posterior intervals.}

\item{control_matches}{Number of matching control markets to use in the analysis}
}
\description{
Given a test market, analyze the impact of an intervention
\code{inference} Analyzes the causal impact of an intervention using the CausalImpact package, given a test market and a matched_market object from the best_matches function.
The function returns an object of type "market_inference" which contains the estimated impact of the intervention (absolute and relative).
}
\examples{
library(MarketMatching)
##-----------------------------------------------------------------------
## Analyze causal impact of a made-up weather intervention in Copenhagen
## Since this is weather data this is a meaningless example and we should
## expect no causal impact. This is just to demo the function.
##-----------------------------------------------------------------------
data(weather, package="MarketMatching")
mm <- best_matches(data=weather, id="Area",
                   date_variable="Date",
                   matching_variable="Mean_TemperatureF",
                   parallel=FALSE,
                   start_match_period="2014-01-01",
                   end_match_period="2014-10-01")
library(CausalImpact)
results <- inference(matched_markets=mm,
                     test_market="CPH",
                     end_post_period="2015-12-15",
                     prior_level_sd=0.002)
}

