% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weather}
\alias{weather}
\title{Weather dataset}
\format{
A time series dataset with 6,935 rows and 3 variables (19 airports and 365 days):
\itemize{
  \item Area: Airport code
  \item Date: Date
  \item Mean_TemperatureF: Average temperature
}
}
\usage{
weather
}
\description{
The data was extracted using the weatherData package
It contains average temperature readings for 19 airports for 2014.
}
\keyword{datasets}
