% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weather}
\alias{weather}
\title{Weather dataset}
\format{A time series dataset with 7,300 rows and 3 variables (20 airports and 356 days):
\itemize{
  \item Area: Airport code
  \item Date: Date
  \item Mean_TemperatureF: Average temperature
}}
\usage{
weather
}
\description{
The data was extracted using the weatherData package
It contains average temperature readings for 20 market for 2014 (365 days)
}
\keyword{datasets}

